Inainte de retestare

+++ api/comments-microservice/src/main/java/com/socialsync/commentsmicroservice/api/CommentsController.java
@@ -1,20 +1,25 @@
 package com.socialsync.commentsmicroservice.api;
 
+import com.socialsync.commentsmicroservice.pojo.AuthorizedInfo;
 import com.socialsync.commentsmicroservice.pojo.Comment;
+import com.socialsync.commentsmicroservice.service.AuthorizationService;
 import com.socialsync.commentsmicroservice.service.CommentsService;
 import com.socialsync.commentsmicroservice.util.exceptions.CommentNotFound;
+import com.socialsync.commentsmicroservice.util.exceptions.UnauthorizedException;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("api/v1/comments")
 @AllArgsConstructor
 public class CommentsController {
     private final CommentsService commentsService;
+    private final AuthorizationService authorizationService;
 
     @GetMapping
     public ResponseEntity<HashMap<String, Comment>> fetchAllComments() {
@@ -31,29 +36,57 @@ public class CommentsController {
     }
 
     @PostMapping
-    public ResponseEntity<Comment> addComment(@RequestBody Comment comment) {
-        commentsService.addComment(comment);
-
-        return new ResponseEntity<>(comment, HttpStatus.CREATED);
+    public ResponseEntity<?> addComment(@RequestHeader("Authorization") String authorizationHeader, @RequestBody Comment comment) {
+        try {
+            AuthorizedInfo authorizedInfo = authorizationService.authorized(authorizationHeader);
+            comment.setCreatorId(authorizedInfo.getId());
+            comment.setId(null);
+            commentsService.addComment(comment);
+            return new ResponseEntity<>(comment, HttpStatus.CREATED);
+        } catch (UnauthorizedException ex) {
+            return new ResponseEntity<>(ex.getMessage(), HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Comment> updateComment(@PathVariable String id, @RequestBody Comment comment) {
+    public ResponseEntity<?> updateComment(@RequestHeader("Authorization") String authorizationHeader, @PathVariable String id, @RequestBody Comment comment) {
+        AuthorizedInfo authorizedInfo = null;
         try {
-            commentsService.updateComment(id, comment);
-            return new ResponseEntity<>(comment, HttpStatus.OK);
-        } catch (CommentNotFound ex) {
+            authorizedInfo = authorizationService.authorized(authorizationHeader);
+            Comment comment1 = commentsService.fetchCommentById(id);
+            if (Objects.equals(comment1.getCreatorId(), authorizedInfo.getId())) {
+                comment.setCreatorId(authorizedInfo.getId());
+                commentsService.updateComment(id, comment);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>("Este topicul altui utilizator!", HttpStatus.UNAUTHORIZED);
+            }
+        } catch (UnauthorizedException ex) {
+            return new ResponseEntity<>(ex.getMessage(), HttpStatus.UNAUTHORIZED);
+        } catch (CommentNotFound topicNotFound) {
+            System.out.println("Topicul nu exista il creez");
+            comment.setCreatorId(authorizedInfo.getId());
+            comment.setId(null);
+            commentsService.addComment(comment);
             return new ResponseEntity<>(comment, HttpStatus.CREATED);
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteComment(@PathVariable String id) {
+    public ResponseEntity<String> deleteComment(@RequestHeader("Authorization") String authorizationHeader, @PathVariable String id) {
         try {
-            commentsService.deleteComment(id);
-            return new ResponseEntity<>("", HttpStatus.NO_CONTENT);
-        } catch (CommentNotFound ex) {
-            return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+            AuthorizedInfo authorizedInfo = authorizationService.authorized(authorizationHeader);
+            Comment comment = commentsService.fetchCommentById(id);
+            if (Objects.equals(authorizedInfo.getRole(), "admin") || Objects.equals(authorizedInfo.getId(), comment.getCreatorId())) {
+                commentsService.deleteComment(id);
+                return new ResponseEntity<>("", HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>("Comentatiul nu poate fi sters decat de admin sau utilizatorul care la creat!", HttpStatus.UNAUTHORIZED);
+            }
+        } catch (UnauthorizedException ex) {
+            return new ResponseEntity<>(ex.getMessage(), HttpStatus.UNAUTHORIZED);
+        } catch (CommentNotFound e) {
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
         }
     }
 }
